+++ script/RegisterReservedNames.s.sol
@@ -12,34 +12,34 @@ contract RegisterReservedNames is Script {
     function run() public {
         // Get the controller address from the environment
         address payable controllerAddress = payable(vm.envAddress("CONTROLLER_ADDRESS"));
-        
+
         // Get the private key from the environment
         uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
-        
+
         // Start broadcasting transactions
         vm.startBroadcast(deployerPrivateKey);
 
         // Get the deployer address (msg.sender)
         address deployer = vm.addr(deployerPrivateKey);
         console.log("Registering reserved test names with address:", deployer);
-        
+
         // Get the controller contract
         DotHypeController controller = DotHypeController(controllerAddress);
-        
+
         // Registration duration - 1 year
         uint256 duration = 365 days;
         console.log("Registration duration:", duration, "seconds (1 year)");
-        
+
         // Test names to register - we'll try all 10, but only the ones reserved for msg.sender will succeed
-        for (uint i = 0; i < 10; i++) {
+        for (uint256 i = 0; i < 10; i++) {
             string memory name = string(abi.encodePacked("test", vm.toString(i + 1)));
-            
+
             (bool isReserved, address reservedFor) = controller.checkReservation(name);
-            
+
             if (isReserved) {
                 if (reservedFor == deployer) {
                     console.log("Attempting to register reserved name:", name);
-                    
+
                     try controller.registerReserved(name, duration) returns (uint256 tokenId, uint256 expiry) {
                         console.log("SUCCESS! Registered", name, "- Token ID:", tokenId);
                     } catch Error(string memory reason) {
@@ -54,10 +54,10 @@ contract RegisterReservedNames is Script {
                 console.log("Skipping", name, "- Not reserved");
             }
         }
-        
+
         // Stop broadcasting transactions
         vm.stopBroadcast();
-        
+
         console.log("Registration of reserved names complete!");
     }
-} 
+}
